energy/requirements.txt
pyserial
influxdb
paho-mqtt